+++ src/components/Member/MemberHeader.jsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
-import LeftButton from './LeftButton';
-import Title from './Title';
+import LeftButton from '../Header/LeftButton';
+import Title from '../Header/Title';
 
 const StyledHeader = styled.div`
   display: flex;
@@ -16,12 +16,12 @@ const TitleWrapper = styled.div`
 `;
 
 //  해당 함수에 온클릭 이벤트 작성
-const onClickButton = () => {};
+const onClickLeftButton = () => {};
 
 const UserHeader = () => {
   return (
     <StyledHeader>
-      <LeftButton onClick={onClickButton} />
+      <LeftButton onClick={onClickLeftButton} />
       <TitleWrapper>
         <Title text="멤버" />
       </TitleWrapper>
